requirements.txt
aiohttp==3.6.2
appdirs==1.4.4
async-timeout==3.0.1
attrs==19.3.0
binance==0.3
certifi==2020.6.20
chardet==3.0.4
discord.py==1.3.3
distlib==0.3.1
get==0.0.0
idna==2.10
importlib==1.0.4
importlib2==3.5.0.2
multidict==4.7.6
post==0.0.0
psycopg2-binary==2.8.5
python-dateutil==2.8.1
request==0.0.0
requests==2.24.0
schedule==0.6.0
six==1.15.0
urllib3==1.25.9
websockets==8.1
yarl==1.4.2
